Upraded Project to URP
I Upgraded the Project to the universal render Pipeline and 2D renderer, wich features 2D lighting and better Post Proessing.

The upgrading process was completely automated( except for the particle systems wich needet to be manually switched to a different material.

the workflow will be completely the same except for the 2D lighting(if you want a scene without lighting just craete a global light 2D)